% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deaths_april.R
\docType{data}
\name{deaths_april}
\alias{deaths_april}
\title{Deaths in April}
\format{
An object of class \code{matrix} with 16 rows and 2 columns.
}
\usage{
deaths_april
}
\description{
Number of confirmed deaths from covid-19 in Sweden during the first
16 days of april.
}
\keyword{datasets}
