% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_s_cov.R
\name{f_s_cov}
\alias{f_s_cov}
\title{f_* covariance function}
\usage{
f_s_cov(w_p, k_ss, k_sx, k_xs, k_xx, x)
}
\arguments{
\item{w_p}{A draw from the posterior of w}

\item{k_ss}{K(X*,X*) matrix}

\item{k_sx}{K(X*, X) matrix}

\item{k_xs}{K(X, X*) matrix}

\item{k_xx}{K(X,X) matrix (from the training data)}

\item{x}{Matrix of training covariates}
}
\description{
Calculates the posterior covariance matrix for test input X_* for a
given draw from the posterior of w.
}
