% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_s_mean.R
\name{f_s_mean}
\alias{f_s_mean}
\title{f_* mean function}
\usage{
f_s_mean(w_p, k_xx, k_sx, y, x)
}
\arguments{
\item{w_p}{A draw from the posterior of w}

\item{k_xx}{K(X,X) matrix (from the training data)}

\item{k_sx}{K(X*, X) matrix}

\item{y}{Observed outcomes}

\item{x}{Matrix of training covariates}
}
\description{
Calculates the posterior mean for test input X_* for a given draw
from the posterior of w.
}
