% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GalaxyData.R
\docType{data}
\name{GalaxyData}
\alias{GalaxyData}
\title{Deaths in April}
\format{
An object of class \code{data.frame} with 82 rows and 1 columns.
}
\usage{
GalaxyData
}
\description{
Number of confirmed deaths from covid-19 in Sweden during the first
16 days of april.
}
\keyword{datasets}
