% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidarData.R
\docType{data}
\name{lidarData}
\alias{lidarData}
\title{Arabidopsis QTL data on gravitropism}
\format{
An object of class \code{data.frame} with 221 rows and 2 columns.
}
\usage{
lidarData
}
\description{
Arabidopsis QTL data on gravitropism
}
\keyword{datasets}
