% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metropolis_sampler.R
\name{metropolis_sampler}
\alias{metropolis_sampler}
\title{A Metropolis sampler}
\usage{
metropolis_sampler(startVals, y, X, iter, burnin)
}
\arguments{
\item{startVals}{Starting values for the parameters}

\item{y}{Response variable}

\item{X}{matrix of covariates}

\item{iter}{Number of iterations the sampler will run}

\item{burnin}{Number of iterations that will be discarded}
}
\description{
This function implements a simple Metropolis MC function
}
