% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newtons_method.R
\name{newtons_method}
\alias{newtons_method}
\title{Newton's method for finding optima}
\usage{
newtons_method(y, k_xx, precision = 0.05)
}
\arguments{
\item{y}{Vector of outcome variables.}

\item{k_xx}{Covariance matrix of the GP.}

\item{precision}{Measured by the squared distance between two consequtive
f vectors. Defaults to 0.05.}
}
\description{
A simple implementation of Newton's method to find posterior mode.
For use with the Laplace approximation in exercise 2, lab1.
}
