% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_kern.R
\name{se_kern}
\alias{se_kern}
\title{Squared exponential kernel}
\usage{
se_kern(x_1, x_2, length_scale, sigma_f)
}
\arguments{
\item{x_1}{First design matrix}

\item{x_2}{Second design matrix}

\item{length_scale}{The length scale}

\item{sigma_f}{Some kind of noise}
}
\description{
An implementation of the squared exponential kernel, K(x_1, x_2).
Note that x_1 = x_2 is valid.
}
