% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_kernel.R
\name{SE_kernel}
\alias{SE_kernel}
\title{Squared exponential kernel}
\usage{
SE_kernel(X, lengthScale, sigma_f)
}
\arguments{
\item{X}{The design matrix}

\item{lengthScale}{The length scale}

\item{sigma_f}{Some kind of noise}
}
\description{
Squared exponential kernel
}
