% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w_post.R
\name{w_post}
\alias{w_post}
\title{Posterior distribution for w}
\usage{
w_post(w, dist_args)
}
\arguments{
\item{w}{Values of the paramaters}

\item{dist_args}{Vector of function arguments, in order they are
y, x, length_scale, sigma_f, and tau2.
Written as list(y, x, length_scale, sigma_f, tau2).}
}
\description{
A slightly longer description here.
}
