% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidarData.R
\docType{data}
\name{lidarData}
\alias{lidarData}
\title{LIDAR data}
\format{
An object of class \code{data.frame} with 221 rows and 2 columns.
}
\usage{
lidarData
}
\description{
LIDAR data for use with exercise 1 and 3 in the first lab.
}
\keyword{datasets}
