% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metropolis_sampler.R
\name{metropolis_sampler}
\alias{metropolis_sampler}
\title{Metropolis MC sampler}
\usage{
metropolis_sampler(dist, dist_args, start_vals, iter, burnin)
}
\arguments{
\item{dist}{Distribution to draw from}

\item{dist_args}{List of arguments needed for the posterior, in this
case it should be list(y = responde, x = covariates, length_scale, sigma_f, tau2)}

\item{start_vals}{Starting values for the parameters}

\item{iter}{Number of iterations the sampler will run}

\item{burnin}{Number of iterations that will be discarded}
}
\description{
This function implements a simple Metropolis MCMC algorithm.
}
